Extract EAGLE3 speculative path
